pyproject.toml
[tool.poetry]
name = "traintime"
version = "0.1.0"
description = "Long Island Rail Road GTFS Parser"
authors = ["Steven Kreuzer <skreuzer@FreeBSD.org>"]
license = "BSD-2-Clause"
readme = "README.md"
repository = "https://github.com/skreuzer/traintime"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
traintime = "traintime.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
